+++ src/char_sql/int_party.h
@@ -4,6 +4,6 @@
 int inter_party_parse_frommap(int fd);
 int inter_party_sql_init();
 void inter_party_sql_final();
-int inter_party_leave(unsigned long party_id,unsigned long account_id);
+int inter_party_leave(uint32 party_id, uint32 account_id);
 
 #endif
