+++ src/char_sql/int_storage.h
@@ -3,8 +3,8 @@
 
 int inter_storage_sql_init();
 void inter_storage_sql_final();
-int inter_storage_delete(unsigned long account_id);
-int inter_guild_storage_delete(unsigned long guild_id);
+int inter_storage_delete(uint32 account_id);
+int inter_guild_storage_delete(uint32 guild_id);
 
 int inter_storage_parse_frommap(int fd);
